added tetrisAngelFinal.asm
this is my final version of the MIPS tetris game, I added end game condition and fixed dropping bugs where blocks would go into each other and not collide and for some reason line would not drop all the way down when it could so i made a special case for line. This version is only missing rotate and clearing rows but should work properly with everything else.